Búsqueda de colores

+++ to-do-pomodoro/quasar-to-do/src/pages/ToDo.vue
@@ -1,53 +1,3 @@
-<template>
-  <div class="todo-box">
-    <div class="add-task-container">
-      <q-input
-        v-model="nuevaTarea"
-        @keyup.enter="agregarTarea"
-        class="add-task-input"
-        square
-        filled
-        placeholder="Añadir tarea"
-        dense
-      >
-        <template v-slot:append>
-          <q-btn @click="agregarTarea" round dense flat icon="add" />
-        </template>
-      </q-input>
-    </div>
-
-    <div class="task-list">
-      <q-list separator bordered>
-        <q-item v-for="(tarea, indice) in tareas" :key="indice" v-ripple>
-          <q-item-section avatar @click="alternarTarea(tarea)">
-            <q-checkbox v-model="tarea.hecha" color="primary" />
-          </q-item-section>
-          <q-item-section @click="alternarTarea(tarea)">
-            <q-item-label :class="{ hecha: tarea.hecha }">{{
-              tarea.titulo
-            }}</q-item-label>
-          </q-item-section>
-          <q-item-section v-if="tarea.hecha" side>
-            <q-btn
-              @click.stop="eliminarTarea(tarea.id)"
-              flat
-              round
-              dense
-              color="primary"
-              icon="delete"
-            />
-          </q-item-section>
-        </q-item>
-      </q-list>
-    </div>
-
-    <div v-if="!tareas.length" class="no-tasks">
-      <q-icon name="check" size="100px" color="primary" class="no-tasks-icon" />
-      <div class="no-tasks-message">No hay tareas</div>
-    </div>
-  </div>
-</template>
-
 <script>
 import { ref } from "vue";
 import axios from "axios";
@@ -66,21 +16,16 @@ export default {
     const tareas = ref([]);
 
     const eliminarTarea = (tareaId) => {
-      dialog({
-        title: "Confirmar",
-        message: "¿Estás seguro/a?",
-        cancel: true,
-        persistent: true,
-      }).onOk(() => {
+      if (window.confirm("¿Estás seguro/a?")) {
         deleteTarea(tareaId)
           .then(() => {
             tareas.value = tareas.value.filter((t) => t.id !== tareaId);
-            notify("Tarea eliminada");
+            console.log("Tarea eliminada");
           })
           .catch((error) => {
             console.error(error);
           });
-      });
+      }
     };
 
     const agregarTarea = () => {
@@ -97,7 +42,7 @@ export default {
             console.error(error);
           });
       } else {
-        notify("La tarea está vacía");
+        console.log("La tarea está vacía");
       }
     };
 
@@ -132,6 +77,61 @@ export default {
 };
 </script>
 
+<template>
+  <div class="todo-box">
+    <div class="add-task-container">
+      <q-input
+        v-model="nuevaTarea"
+        @keyup.enter="agregarTarea"
+        class="add-task-input"
+        square
+        filled
+        placeholder="Añadir tarea"
+        dense
+      >
+        <template v-slot:append>
+          <q-btn @click="agregarTarea" round dense flat icon="add" />
+        </template>
+      </q-input>
+    </div>
+
+    <div class="task-list">
+      <q-list separator bordered>
+        <q-item v-for="(tarea, indice) in tareas" :key="indice" v-ripple>
+          <q-item-section avatar @click="alternarTarea(tarea)">
+            <q-checkbox v-model="tarea.hecha" color="red" />
+          </q-item-section>
+          <q-item-section @click="alternarTarea(tarea)">
+            <q-item-label :class="{ hecha: tarea.hecha }">{{
+              tarea.titulo
+            }}</q-item-label>
+          </q-item-section>
+          <q-item-section v-if="tarea.hecha" side>
+            <q-btn
+              @click.stop="eliminarTarea(tarea.id)"
+              flat
+              round
+              dense
+              color="red"
+              icon="delete"
+            />
+          </q-item-section>
+        </q-item>
+      </q-list>
+    </div>
+
+    <div v-if="!tareas.length" class="no-tasks">
+      <q-icon
+        name="check"
+        size="100px"
+        color="$principal"
+        class="no-tasks-icon"
+      />
+      <div class="no-tasks-message">No hay tareas</div>
+    </div>
+  </div>
+</template>
+
 <style lang="scss">
 html,
 body {
@@ -140,10 +140,11 @@ body {
 }
 
 .todo-box {
-  background-color: lightcoral;
+  background-color: $principal;
   width: 100%;
   height: 90vh;
 }
+
 .no-tasks {
   text-align: center;
   margin-top: 25%;
